## 02/09/19 10:52 to 11:16, Children in Need Funding Application +EQT It's a midnight deadline and I've put a call out for reviewers
Some back of the envelope calculations, based on how it worked on the
pilot suggests that the leader needs 16 hours per school to get
something started and then come back for research collection, including
setting up and transport and faff.  It's actually less for part time
staff presuming that they can fit in for when schools want them. Given
the proper school year we're thinking that a full time facilitator can
do 50 schools and 0.5 can do 30, with 20 for 0.3.

Fitting a reasonable project into £10k a year is going to be hard -
we're going to be absorbing a lot of this into the core costs if it
comes off.  But we do believe in it as a project.

I've cut the project manager role right down. It's come out at 999.30
which I'm willing to take.

So we're hoping for 40 schools over the two year project. Or 60 over
three.

## 02/09/19 11:24 to 11:49, Worked out an answer and quickly blogged it.
+EQT
## 02/09/19 11:49 to 12:29, writing until lunch.+EQT
## 02/09/19 13:18 to 13:50, Doing the finance +EQT
## 02/09/19 16:18 to 16:43, I'm back for the second, more complete,
fufilled version of the application. +EQT
## 02/09/19 21:45 to 22:28, Application Submitted +EQT

## 05/09/19 09:45 to 10:42, CM2019 presentation Prep +EQT