## 11/09/19 09:28 to 11:43, Sorting out the old git projects in Dropbox +EQT Good sick day activity. There were 12G of them (with some personal projects mixed in) The main query was 'what to do about desktop tracking' which I've moved to a different bit of dropbox'
This eventually went down to less than a gig

Obviously this doesn't end up being 'done' until I take it out of dropbox as well.

## 12/09/19 11:58 to 12:40, working on meta
When I do the rebase with -i there are 1372 commits? But only 600 lines in the log.csv?

## 12/09/19 13:22 to 14:47
Notably an enomouse percentage of my todo list is (E) I wonder if there is a better way of doing that.

## 12/09/19 14:47 to 14:57, More rewriting.
## 12/09/19 14:58 to 16:00, Power hour +PersonalProjects

## 12/09/19 20:56, Hello,