## 17/09/19 01:00 to 04:35, +Bed ## 17/09/19 04:59,  Morning!
## 17/09/19 05:23 to 05:52, (A due:2019-10-01 Trustees, Funder) Sprint on SupCy.todo.txt +EQT

## 17/09/19 05:53 to 06:38, Email again  +EQT
Time flew on that one, v.strange, no phone or anything.

## 17/09/19 06:38 to 09:18, +Family (and small work email)

## 17/09/19 09:18 to 10:56, Hmmm,
There was a lot of work in this (fixing bugs, contacting people about things) but there was also a lot of faff and I don't think I can count it. :(

So far today, I am not ahead in terms of work... :(