## 12/11/19 06:22 to 07:38, +EQT accounts I've been working on the accounts, I'm checking that the accounts from 2017-2018 match up with Quickbooks before we start the next account. * Assets Match * Cash in bank matches * Debtors doesn't match
This has been good. I've invoiced a couple of camps I should have done
long ago, and corrected some issues, there are still some big bits to
do.

## 12/11/19 07:38 to 09:19, +Family

## 12/11/19 09:19 to 10:47, +EQT accounts
* Working out how to account for all the SHINE money, and match it up
* with the existing school invoices and so on.
* I've sent a few reminder emails to people with outstanding invoices
* I've adjusted the dates on several THINGs
* I've finally processed the Kings Purchase order
* I lost 1200 in the system and then found it again
* I've found a new and better way to see the open invoices.

## 12/11/19 10:47
Whoop! Debetors is correct! Finally!

## 12/11/19 10:47 to 12:35, +EQT Accounts, now dealing with creditors
Okay, turns out there are several problems with the creditors.

1. Because there aren't any bills at that time, the expenses (book camps
and so on) don't show up in Quickbooks. This is about 6k

2. Deferred income doens't make sense to me - Comic Relief is out by 474
- the exact amount of the Communication Matters Conference (which
  admittedly, takes place in the next financial year) But Nesta is far
lower than I have it, and there are entries for Media Wall and Awesome
that I wouldn't have put in.

I've written to the accountant