## 12/12/19 10:21  to 11:11, Writing up ACCA course and dealing with Cover artists +EQT ## 12/12/19 11:12 to 11:30, Writing up another article. +EQT
We need to talk about prices.

We haven't had the prices on the website so far because they vary and we
have a lot of ways of delivering the camps that make a difference.
That's the wrong way of doing things, what we *should* have on the
website is an explanation.

So a White Water Writers camp generally costs about £1,400 to run. Most
of that is staff costs, most of the rest is the cost of the books
themselves.  Normally this is what we charge schools for a camp.

If your school is a long way away from any of our existing leaders, I'll
say so and add in the cost of putting a leader in a hotel for the week.

If you have multiple camps at once (we've done up to nine in a single
week in one school), then I don't need as many staff (Camp need a 10:1
writer:leader ratio on Monday and Tuesday, then 20:1 from Wednesday)
that makes it a little cheaper - around £3,100 for two.

Most importantly, if we've trained up some of your staff, then you can
run it on your own, and you'd only pay the services fee, which is £700.
Currently the vast majority of our books are delivered this way - with
local staff who don't actually work for us.

Generally if you have a member of staff stay with us for one full camp,
and they feel comfortable with it, we'll let them lead their own. (There
are various 'in-between' setups, which are generally 'Joe turns up for
the Monday and Tuesday to make sure everything is going His Way').

So the prices can be very different, depending on the costs.

## 12/12/19 11:30 to 11:45, +Family call
## 12/12/19 11:45, checked gestal meeting rooms  +EQT
## 12/12/19 11:45 to 13:36, Working on Rob's Recommendations to get them ready for Trustees  +EQT
## 12/12/19 13:37 to 13:52, Working on https://github.com/eQualityTime/Public/issues/96.  +EQT

## 12/12/19 13:56 to 14:16, Blogging and sharing everywhere. Takes 20 minutes. That's too much I think. +EQT

## 12/12/19 14:16 to 14:23, Writing up notes +EQT

I'm actually not that behind. I could totally put this in. In fact I
will. Time to properly write up and shut down.