# Monday Did a really good stream and sorted out the Wheel of Life for the next sprint, also *finally* ticked off the git directory
## 27/01/20 07:57 to 08:17,  New stream, tidy up first  +EQT
* Clearing out files, closing windows, making sure applications are the
* right ones.
* Starting on 33 clicks
* Processed all the bits of paper/reminders (some where personal)

## 27/01/20 08:18 to 08:47, Review +PersonalProject s in Github
## 27/01/20 08:47 to 09:10, Review +EQT projects in Github

## 27/01/20 09:10, Okay, all the issues have been given a general look
at.  Let's fix the notes.

## 27/01/20 09:10 to 11:37, Working on Wheel of life +PersonalProject
It was the end of a 'wheel of life' sprint yesterday so it was time to
add up my scores. They were extremely dismal - averaged about 2.5 out of
ten.

I've adjusted various goals for the next three months, including putting
in actual goals for the three categories that didn't get them last time.
I've also done a certain amount of general tidying up of the sheets,
some brainstorming on what I really want (actually a lot of this was
done prior, I just now had the energy to do it *properly*.

## 27/01/20 16:30 to 17:23, clearing out the git folder +EQT