# Wednesday! This was a good day - got lots done, got a shower, saw a friend
## 29/01/20 10:20 to 12:25, Starting Stream on Strategic Plan +EQT
Today I want to write out my suggestion for eQuality Time's Stratergy
for the next ten years.

What would we do if we had to do this, *this* year? (as an exercise)

* Hire a fundraiser
* Advertise OVF a lot using the free adwords
* Get research done on TMC immediately so we can count them.
* See our other projects in this light 'okay, this is cool, but what is
* the cost of the alternative?'

## 29/01/20 12:25 to 13:57, sauna, shower, quick browse of shops

## 29/01/20 13:57 to 14:37, Writing more stragic plan +EQT

## 29/01/20 15:22 to 15:58, working on strategic plan more +EQT
* Properly falling asleep at the end.

## 29/01/20 16:00 to 21:51, +Family time
Had a visitor, that was nice. Healthy Dinner, played a board game.

## 29/01/20 21:51 to 23:04, Unexpected OVF bug +EQT
Got a message from J at Optikey that there was a problem with my code.
Spent some time making it better (not doing well at that) and got
reinspired to play with some code.