## 10/02/20 09:30 to 09:59, Attempting to get a command line twitter client working Will have to upvote this: https://apple.stackexchange.com/questions/126397/run-sudo-from-non-admin-account Eventually realised tweet deck without columns did most of what I needed.
## 10/02/20 10:00 to 10:04, Now working on todo list +EQT
## 10/02/20 10:04 to 10:30, Working on https://github.com/eQualityTime/Public/issues/104  +EQT
## 10/02/20 10:30 to 10:48, +EQT
* Created a repeating event to tell trustees about changes to
https://github.com/orgs/eQualityTime/projects/3
* Posted to WWW Facebook page
* Dealt with window cleaner
* Reviewed todo list

## 10/02/20 10:48 to 11:12, CPD on vim. +EQT

* I've leanred that <CR> puts you on the first characte rof the next
line, and that the auto tabbing is different in places , and that cronol
o puts a normal command in in insert mode

## 10/02/20 11:13 to 11:32, Working on other todo list
Okay, so full-accounts-algorithm, isn't comeinth I can do on a public
steam. Let me pop onto the other todo for a while and see if I can do a
few minutes worth of work.

## 10/02/20 13:05 to 13:24, Office clearing and cleaning +EQT
Office cleaning I think