# Tuesday On Monday someone asked my advice about freelancing and I told him that everything good that had happened to me in buisness could be traced back to meeting someone at an event so he should be going to events of any type regularly. I then had to admit that I’d hardly left the house for a couple of years because I absolutely hated the idea of networking.
I was shamed by that so this morning I picked a interesting looking (and
free) event off eventbright and wandered in.

It was run by a man who I had never heard of but who turned out to be a
guru of some sort. People in the audience had been going to his events
for 20 years and he preached a gospel that while massively different
from a lot of what I believe, was a least fun and well delivered; and I
had a whale of a time chatting to various people afterwards (I
networked!) and got some new perspectives on things and clarified
thoughts I’d been on the edge of having for years.

That was good. Also: turns out there is a university in the middle of
Regent’s Park. That was news to me.

In need of some significant introvert time I then headed over for a
cinema trip with a friend- we saw 1917 and I sobbed during at least five
separate parts. Exhausting. Awesome film. Really something: a real slap
in the face in a ‘you think you have problems’ way. There was also
flapjack and salad that had questionable legal status.

One catch up with a close friend turned into another when one sprang
into view at Euston station. A proper validating and fun conversation
later, I went to a support group, which was cool, although not how I
remembered it, and I wanted to tell you that because it’s the sort of
thing I want people to be able to tell other people.

## 11/02/20 09:40 to 14:25, +EQT buisness networking
* I'm charging five minutes of a much longer conversation with FB to
* this

## 11/02/20 14:25 to 23:00, out with friends in London