## 15/04/20 06:01 "(C) Put nails in the wall for more pictures" Bit loud, but can rewrite and have
## 15/04/20 06:03 to 06:51, (C) Rememorise Kat phone number
(Plus some family interuption)

## 15/04/20 06:52
Got to the garage to get the THING

## 15/04/20 07:05 to 09:13, Working on https://github.com/joereddington/Projects-Public/issues/35 +PersonalProjects

## 15/04/20 09:21, Been tidying up todo list,

Tomorrow is email, and the day after is new list

## 15/04/20 18:00
So the thing I'm currently wondering is: where does watcher fit in? Is it a seperate program? Not really because it needs vim to work in a particular way. Can it be part of watson? No because watson is just a mess.

should it be under version control: yes, everything else is.
* Should it be part of vimrc?
  * with a soft link for the web directory.

## 15/04/20 18:04, What's on the list
(C) Properly repair the memory palace
Start with the periodic table. o

## 16/04/20 06:34 to 07:11, Email