## 28/04/20 08:38 to 09:50, Working on +EQT There's a problem with insurance to investigate * I started writing up the TRUST NOBODY briefing note
Feeling a bit panicked, but only because I'm behind on Yvonne's STUFF. Accounts should be fine, prison book should be fine, insurance needs work. I need to do a revision on the grant and write up a bunch of STUFF for it.

## 28/04/20 19:45 to 20:45, writing long email to trustees +EQT

## 29/04/20 06:29 to 11:30, +EQT
Delivering training and also working on the grant proposal.

## 29/04/20 15:00 to 15:29, +EQT working on grant
## 29/04/20 19:30 to 21:05, +EQT meeting and grant