## 08/05/20 05:53 Okay, blogging.
## 08/05/20 06:00 to 08:13, Website work.
A successful morning working on: https://github.com/joereddington/Projects-Private/issues/107

It's capable of being shown to others and works fine locally and on github. I've learned quite a lot and put it in the commit notes where I can.

I've created issues, and closed issues, and because we are fileless I've deleted it again.