## 18/07/20 05:12 to 05:18, Kat Joe Reminders list
## 18/07/20 05:18 to 06:46, Touching every task
Start: 1 minutes 28 seconds for a washington run thorugh
06:09: 1 minutes 49 seconds.

Stopped at 06:39 because I realised I wasn't going to be able to complete it *and* write it up.

It was fun, I cleared a few tasks off the list (72-54), and I'm feeling more on top of it. There were about 10 low-priority tasks I didn't touch but that's okay

- 18/07/20 smallest next step: Ask Kat about giving blood on Monday.