## 18/07/20 05:12 to 05:18, Kat Joe Reminders list
## 18/07/20 05:18 to 06:46, Washington run through
Start: 1 minutes 28 seconds for a washington run thorugh
06:09: 1 minutes 49 seconds.

Stopped at 06:39 because I realised I wasn't going to be able to complete it *and* write it up.

It was fun, I cleared a few tasks off the list (72-54), and I'm feeling more on top of it. There were about 10 low-priority tasks I didn't touch but that's okay

- 18/07/20 smallest next step: Ask Kat about giving blood on Monday.

## 20/07/20 04:20 to 04:48, gratitude
Morning! Doing gratitude, but need a photograph.

## 22/07/20 11:06 to 11:14, (C) get glue from garage for Nova's room  on: https://github.com/joereddington/Projects-Private/issues/149 +PersonalProjects
Let's look at the magenets

## 22/07/20 11:14 to 11:32, Getting to the bottom of some of the tasks

## 23/07/20 11:08 to 11:24, Working on the todo list
I found that I could have the subtasks I want with these commands:
g/*/-1j   (to gather together all of the subtasks)

%s/*/ */g (to open them up again)