## 05/08/20 06:00 to 07:28, * The solution to yesterday's question turned out to be 'write it in two boxes' * Wrote about 1500 words on the third question. Now for the next. * Think I might acutally have a static url... https://www.youtube.com/channel/UCLOmaOSJYS1FVBQXLIGf0QQ/live * This has ended the quality section, so I should check to make sure that I think I've done the things I was meant to.   * "if you have demonstrated that your project is likely to achieve its ambition"   * "whether the artists/organisations involved are high quality in the context you are working in" * Want to go back and put in the bit about artist recruitment   * Done.   * I've also slightly strengthened how confident we are.
Next section.
* Read the section
* Do I want to work backwards through the questions, so they have a narrative.

## 05/08/20 08:19 to 09:36, Finance questions
Okay, so here is the current state of play.

I have written, from scratch, four new answers to Arts Council Questions, and copy and pasted one over. There are seven remaining questions to write answers to as well as writing a audience development plan from scratch and adjusting the budget (carefully!)

I have a couple of guides to work from for the questions. So far it's clear that I'm writing very different answers to the questions than I thought I would, which is nice.

## 05/08/20 09:47
The Next action is to talk to Kat about asking for more time. The action before/after that, is to start the audience development plan.

There's already some good smart goals in the audiendece development plan. Actually the next thing to do is to read it in general .

"Audience includes participants and learners"